+++ CHANGELOG.md
@@ -11,6 +11,7 @@ en-US:
 - Add a refresh serial port button in the serial port selection page
 - Add up to four window split screen mode and multiple layout modes
 - Add single/group/all session three modes in the send command window
+- Add the plugin information page to display the plugin website
 - Improve the context menu of URL recognition links
 - Improve the find window to automatically fill in the currently selected text each time it is opened
 - Improve status bar
@@ -26,6 +27,7 @@ zh-CN:
 - 增加刷新串口按钮在选择串口页面中
 - 增加了至多四窗口分屏模式以及多种布局模式
 - 发送命令窗口增加单会话/组会话/全部会话三种模式
+- 插件信息页面增加插件网站主页显示
 - 改进URL识别链接的上下文菜单
 - 改进查找窗口在每次打开时自动填入当前选择的文本
 - 改进状态栏